Keep ring buffer of data from UART
Enables UART data to be saved while no client is connected. This data
is sent to the client when it connects. The ring buffer is rewinded
as far as possible when a client connects, to enable data to be resent
on reconnection.

The ring buffer has optimized get() and put() functions for handling
bytearrays, as iteratively getting/putting a character at a time
turned out to be too slow on a RPi Pico with a 115200 baud UART.